% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.Colcap}
\alias{get.Colcap}
\title{Colcap extraction}
\usage{
get.Colcap()
}
\value{
The dataframe with dates and the Colcap.
}
\description{
Function to extract the Colcaptime serie from the Colombian Central Bank.
}
\examples{
\dontrun{
colcap <- get.Colcap();
}
}
