% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.IBR}
\alias{get.IBR}
\title{IBR extraction}
\usage{
get.IBR(nom = "ON")
}
\arguments{
\item{nom}{The interest rate period. \cr
The periods available are: \cr
\itemize{
 \item ON: Overnight
 \item 1M: 1 Month
 \item 3M: 3 Months
 \item 6M: 6 Months
}
Default is "ON".}
}
\value{
The dataframe with dates and the IBR.
}
\description{
Function to extract the nominal IBR rate time serie from the Colombian Central Bank.
}
\examples{
\dontrun{
ON <- get.IBR("ON");
M1 <- get.IBR("1M");
M3 <- get.IBR("3M");
M6 <- get.IBR("6M");
}
}
