% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.IPC}
\alias{get.IPC}
\title{IPC extraction}
\usage{
get.IPC()
}
\value{
The dataframe with dates and the Colombian CPI.
}
\description{
Function to extract the Colombian CPI time serie from the Colombian Central Bank.
}
\examples{
\dontrun{
IPC <- get.IPC();
}
}
