% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.TasaDesempleoCol}
\alias{get.TasaDesempleoCol}
\title{Colombian Unemployment Rate extraction}
\usage{
get.TasaDesempleoCol()
}
\value{
The dataframe with dates and the Colombian Unemployment Rate.
}
\description{
Function to extract the Colombian unemployment rate time serie from the Colombian Central Bank.
}
\examples{
\dontrun{
unemp <- get.TasaDesempleoCol();
}
}
