% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.TasaIntBanRep}
\alias{get.TasaIntBanRep}
\title{Colombian Central Bank's Policy Rate extraction}
\usage{
get.TasaIntBanRep()
}
\value{
The dataframe with dates and the Policy Rate.
}
\description{
Function to extract the Colombian Central Bank's Policy Rate time serie from the Colombian Central Bank.
}
\examples{
\dontrun{
intRate <- get.TasaIntBanRep();
}
}
