% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapping_pack.R
\name{get.TRM}
\alias{get.TRM}
\title{TRM extraction}
\usage{
get.TRM()
}
\value{
The dataframe with dates and the TRM.
}
\description{
Function to extract the TRM (COP/USD) time serie from the Colombian Financial Supervision Office.
}
\examples{
\dontrun{
TRM <- get.TRM();
}
}
